// src/interfaces/Schedule.ts
export interface Schedule {
  id: string;
  moduleTitle: string;
  lecturerName: string;
  classroomNumber: string;
  floorNumber: string;
  branch: string;
  startTime: string;
  endTime: string;
  isRecurring: boolean;
  dayOfWeek: string;
  date: string;
  createdAt: string;
  updatedAt: string;
} 